G Suite: Move the isGSuiteOrExtraLicenseOrGoogleWorkspace() selector to its own file
