+++ packages/calypso-products/src/is-gsuite-or-extra-license-or-google-workspace.ts
@@ -1,0 +1,15 @@
+import { camelOrSnakeSlug } from './camel-or-snake-slug';
+import {
+	isGoogleWorkspaceProductSlug,
+	isGSuiteOrExtraLicenseProductSlug,
+} from './gsuite-product-slug';
+import type { WithCamelCaseSlug, WithSnakeCaseSlug } from './types';
+
+export function isGSuiteOrExtraLicenseOrGoogleWorkspace(
+	product: WithCamelCaseSlug | WithSnakeCaseSlug
+): boolean {
+	return (
+		isGSuiteOrExtraLicenseProductSlug( camelOrSnakeSlug( product ) ) ||
+		isGoogleWorkspaceProductSlug( camelOrSnakeSlug( product ) )
+	);
+}
