+++ packages/calypso-products/src/product-values.ts
@@ -39,11 +39,11 @@ export {
 	isGoogleWorkspace,
 	isGoogleWorkspaceExtraLicence,
 	isGSuite,
-	isGSuiteOrExtraLicenseOrGoogleWorkspace,
 	isGSuiteOrGoogleWorkspace,
 } from './is-gsuite';
 export { isGoogleWorkspaceMonthly } from './is-google-workspace-monthly';
 export { isGSuiteOrExtraLicense } from './is-gsuite-or-extra-license';
+export { isGSuiteOrExtraLicenseOrGoogleWorkspace } from './is-gsuite-or-extra-license-or-google-workspace';
 export { isJetpackAntiSpam } from './is-jetpack-anti-spam';
 export { isJetpackAntiSpamSlug } from './is-jetpack-anti-spam-slug';
 export { isJetpackBackup } from './is-jetpack-backup';
